Last concurrency fixes
