+++ BioViewer/Metal/Renderer/RenderPasses/Postprocessing/Upscaling/MetalFXUpscaling.swift
@@ -8,7 +8,7 @@
 import Foundation
 import MetalKit
 
-extension ProteinRenderer.MutableState {
+extension MutableState {
         
     func metalFXUpscaling(
         renderer: ProteinRenderer,
